+++ src/main/java/kr/co/imguru/domain/like/entity/LikePost.java
@@ -1,8 +1,8 @@
-package kr.co.imguru.domain.like;
+package kr.co.imguru.domain.like.entity;
 
 import jakarta.persistence.*;
 import kr.co.imguru.domain.member.entity.Member;
-import kr.co.imguru.domain.review.entity.Review;
+import kr.co.imguru.domain.post.entity.Post;
 import kr.co.imguru.global.common.BaseEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -12,11 +12,11 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LikeReview extends BaseEntity {
+public class LikePost extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "like_review_id")
+    @Column(name = "like_post_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -24,13 +24,13 @@ public class LikeReview extends BaseEntity {
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "review_id")
-    private Review review;
+    @JoinColumn(name = "post_id")
+    private Post post;
 
     @Builder
-    public LikeReview(Member member,
-                      Review review) {
+    public LikePost(Member member,
+                    Post post) {
         this.member = member;
-        this.review = review;
+        this.post = post;
     }
 }
