+++ src/main/java/kr/co/imguru/domain/like/entity/LikeReview.java
@@ -1,9 +1,8 @@
-package kr.co.imguru.domain.like;
+package kr.co.imguru.domain.like.entity;
 
 import jakarta.persistence.*;
 import kr.co.imguru.domain.member.entity.Member;
-import kr.co.imguru.domain.post.entity.Post;
-import kr.co.imguru.domain.reply.entity.Reply;
+import kr.co.imguru.domain.review.entity.Review;
 import kr.co.imguru.global.common.BaseEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -13,11 +12,11 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LikeReply extends BaseEntity {
+public class LikeReview extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "like_reply_id")
+    @Column(name = "like_review_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -25,13 +24,13 @@ public class LikeReply extends BaseEntity {
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "reply_id")
-    private Reply reply;
+    @JoinColumn(name = "review_id")
+    private Review review;
 
     @Builder
-    public LikeReply(Member member,
-                     Reply reply) {
+    public LikeReview(Member member,
+                      Review review) {
         this.member = member;
-        this.reply = reply;
+        this.review = review;
     }
 }
